Refactored JS with classes and constants for chat, folder, and settings functionalities
