+++ static/css/styles.css
@@ -77,7 +77,6 @@ p {
 }
 
 .chat-container {
-    /* Adjust height based on your layout */
     margin-left: 10%;
     margin-right: 10%;
     margin-top: 15px;
@@ -86,10 +85,12 @@ p {
     border-radius: 10px;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
     flex: 1;
-    justify-content: center;
+    display: flex;
+    position: relative;
 }
 
 .chat-messages {
+    flex: 1;
     overflow-y: auto;
     padding-bottom: 0px;
     padding-top: 0px;
@@ -145,16 +146,20 @@ p {
     cursor: pointer;
     font-size: 12px;
     color: white;
+    opacity: 0.7;
+    transition: opacity 0.2s;
 }
 
 .action-button-user:hover {
     color: whitesmoke;
     background: none;
+    opacity: 1;
 }
 
 /* Markdown styles */
 .message-content {
     margin-right: 20px;
+    margin-top: 5px;
     font-size: 14px;
 }
 
@@ -200,14 +205,14 @@ p {
     position: relative;
 }
 
-.bot-message {
+.assistant-message {
     background-color: #e9ecef;
     color: black;
     margin-right: 20%;
     margin-left: 0;
 }
 
-.bot-message-line {
+.assistant-message-line {
     font-size: 0.7rem;
     color: #5a6268;
     height: 16px;
@@ -231,6 +236,45 @@ p {
     border-bottom: 1px solid;
     border-bottom-color: #5a6268;
     position: relative;
+}
+
+.scroll-controls {
+    width: 10px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    padding: 10px 5px;
+}
+
+.scroll-icon {
+    width: 20px;
+    height: 20px;
+    background-color: gray;
+    color: #fff;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    transition: background-color 0.3s;
+}
+
+.scroll-icon i {  /* Added this to target the Font Awesome icons */
+    font-size: 12px;  /* Adjust this value to make the arrows smaller */
+}
+
+.scroll-icon:hover {
+    background-color: rgba(0, 0, 0, 0.8);
+}
+
+.top-icon {
+    top: 10px;
+    position: absolute;
+}
+
+.bottom-icon {
+    bottom: 10px;
+    position: absolute;
 }
 
 .chat-input-container {
@@ -607,12 +651,15 @@ button:hover {
     .user-message {
         margin-left: 7%;
     }
-    .bot-message {
+    .assistant-message {
         margin-right: 7%;
     }
     .system-message {
         margin-left: 0%;
         margin-right: 0%;
+    }
+    .copy-code-button {
+        display: block !important; /* Force display on mobile */
     }
 }
 
@@ -625,6 +672,9 @@ button:hover {
         padding-left: 20px;
         padding-right: 20px;
     }    
+    .message-content pre:hover .copy-code-button {
+        display: block;
+    }
 }
 
 .conversation-selector {
@@ -905,11 +955,6 @@ button:hover {
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
 }
 
-/* Make sure the chat messages container has a relative position 
-#chat-messages {
-    position: relative;
-}*/
-
 .loading-spinner::after {
     content: '';
     display: inline-block;
@@ -925,4 +970,42 @@ button:hover {
 @keyframes spin {
     0% { transform: rotate(0deg); }
     100% { transform: rotate(360deg); }
+}
+
+.message-content pre {
+    background-color: #f5f5f5;
+    border-radius: 4px;
+    padding: 1em;
+    overflow-x: auto;
+    position: relative;
+    padding-bottom: 35px; /* Add some space for the button */
+}
+
+.message-content code {
+    font-family: 'Consolas', 'Monaco', 'Courier New', monospace;
+    font-size: 0.9em;
+}
+
+/* For inline code */
+.message-content :not(pre) > code {
+    background-color: rgba(0, 0, 0, 0.05);
+    padding: 0.2em 0.4em;
+    border-radius: 3px;
+}
+
+.copy-code-button {
+    position: absolute;
+    bottom: 5px;
+    right: 5px;
+    padding: 5px 10px;
+    min-width: 10px !important;
+    background: #fff;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    cursor: pointer;
+    display: none; /* Hidden by default */
+}
+
+.copy-code-button:hover {
+    background: #f0f0f0;
 }
