+++ templates/index.html
@@ -10,37 +10,14 @@
         content="width=device-width, height=device-height, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover">
     <title>Chat Application</title>
     <script src="https://cdn.jsdelivr.net/npm/marked@4.3.0/marked.min.js"></script>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css">
+    <!--<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css">-->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
     <link rel="stylesheet" href="/static/css/styles.css">
     <link rel="icon" type="image/png" href="/static/images/logo-favicon.png">
     <link rel="apple-touch-icon" href="/static/images/logo-iphone-tr.png">
-    <script>
-        async function testProtectedEndpoint() {
-            try {
-                const token = localStorage.getItem('token');
-                if (!token) {
-                    throw new Error('No token found');
-                }
-
-                const response = await fetch('/api/users/me', {
-                    headers: {
-                        'Authorization': `Bearer ${token}`
-                    }
-                });
-
-                if (!response.ok) {
-                    throw new Error('Failed to access protected endpoint');
-                }
-
-                const data = await response.json();
-                console.log('Protected endpoint response:', data);
-                alert('Successfully accessed protected endpoint!');
-            } catch (error) {
-                console.error('Error:', error);
-                alert(error.message);
-            }
-        }
-    </script>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/prism/1.24.1/themes/prism.min.css" rel="stylesheet" />
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/prism/1.24.1/prism.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/prism/1.24.1/components/prism-javascript.min.js"></script>
 </head>
 <!-- 
         <h1>Chat Application</h1>
@@ -132,7 +109,6 @@
             </div>-->
         </div>
         <div class="chat-container">
-            <!-- Messages will be dynamically added in below div -->
             <div class="chat-messages" id="chat-messages"></div>
         </div>
         <div class="chat-input-container">
@@ -154,7 +130,7 @@
             </form>
         </div>
     </div>
-    <script src="/static/js/main.js"></script>
+    <script type="module" src="/static/js/main.js"></script>
 </body>
 
 </html>
